feat: makeGame function
