+++ README.md
@@ -18,6 +18,7 @@
 ### DONE
 * [X] 자동차 갯수와 이동 횟수 입력받기
 * [X] 정상적인 입력인지 확인하기
+* [X] 입력값에 맞게 자동차 및 레이스 게임 생성
 * [X] random 값 생성하기
 
 ## 프로그래밍 요구사항
