+++ src/test/java/racingcar/RacingGameTest.java
@@ -5,17 +5,17 @@ import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-public class RacingCarTest {
+public class RacingGameTest {
 
     @Test
     public void validate_음수_입력값() {
-        assertThatThrownBy(() -> RacingCar.validateInput(new int[] {-1, 3}))
+        assertThatThrownBy(() -> RacingGame.validateInput(new int[] {-1, 3}))
                 .isInstanceOf(RuntimeException.class);
     }
 
     @Test
     public void random_범위() {
-        int random = RacingCar.getRandomValue();
+        int random = RacingGame.getRandomValue();
         assertThat(random)
                 .isGreaterThanOrEqualTo(0)
                 .isLessThanOrEqualTo(9);
